fixed the problems i created with switching to int and commented out troublesome postmapping in check in controller

+++ src/main/java/com/launchcode2026/kunce_self_love_bombs/repositories/LoveBombRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LoveBombRepository extends JpaRepository<LoveBomb, Long> {
+public interface LoveBombRepository extends JpaRepository<LoveBomb, Integer> {
 
     List<LoveBomb> findByCategoryIgnoreCase (String category);
     List<LoveBomb> findByMessageContainingIgnoreCase(String keyword);
